// design-system/src/components/filters-group-factory/filters-group-factory.interfaces.ts
type IdentificationType = 'chile-rut';

export type FilterFactoryContentTypeVariant =
  | 'text'
  | 'number'
  | 'email'
  | 'boolean'
  | IdentificationType;

export type ItemId = Item['id'];

/**
 * Defines the raw values for each filter type. This is the shape of data
 * that the filter is responsible for storing, which can be used to
 * restore the filter’s state or persist it (e.g., in local storage).
 */
export type FilterFactoryRawValuesMap = {
  list: {
    pathIds: ItemId[];
    checkedIds: ItemId[];
  };
  string: {
    value: string;
  };
  datepicker: {
    date: string | string[] | null;
  };
};
// ...
export type FilterFactoryRawValues<T extends string = string> = {
  id: T;
} & {
  [K in ContentTypes]: {
    type: K;
    typeVariant?: FilterFactoryContentTypeVariant;
  } & FilterFactoryRawValuesMap[K];
}[ContentTypes];